+++ examples/shorthands.rb
@@ -6,7 +6,7 @@ raise "err" unless k.out == "hello"
 puts k.out
 
 k = Kommando.run_async "$ echo hello"
-sleep 0.1 until k.code == 0  #TODO: k.wait
+k.wait
 
 raise "err" unless k.out == "hello"