+++ spec/kommando/run_opts_spec.rb
@@ -41,8 +41,8 @@ describe Kommando do
         }
         expect { k.run }.not_to output(/\d+ users, load averages:/).to_stdout
 
-        sleep 0.01 until k.code == 0  #TODO: k.wait
-        sleep 0.5
+        k.wait
+        sleep 0.5 # wait for flush..
 
         contents = File.read outfile.path
         expect(contents).to match /\d+ users, load averages:/