+++ spec/kommando/run_spec.rb
@@ -33,11 +33,12 @@ describe Kommando do
       end
 
       it 'calls passes args correctly' do
-        k = Kommando.new "head .rspec"
+        k = Kommando.new "echo 1"
         expect(k.run).to eq true
         expect(k.code).to eq 0
-        sleep 0.5 # TODO: bin/stress 10 100 makes this fail
-        expect(k.out).to eq "--format documentation\r\n--color\r\n"
+        k.wait
+
+        expect(k.out).to eq "1\r\n"
       end
     end